update TestBatchPointGetPartitionForAccessObject case
