+++ planner/core/casetest/partition/testdata/integration_partition_suite_in.json
@@ -256,5 +256,16 @@
       "select * from tlist3 where a in (1,2) and 1 = 1 order by a",
       "select * from tlist3 where a in (1,2) and 1 = 1 order by a desc"
     ]
+  },
+  {
+    "name": "TestBatchPointGetPartitionForAccessObject",
+    "cases": [
+      "explain format='brief' select * from t1 where b in (1, 2)",
+      "explain format='brief' select * from t1 where b in (1, 2, 1)",
+      "explain format='brief' select * from t2 where id in (1, 3)",
+      "explain format='brief' select * from t3 where id in (1, 3)",
+      "explain format='brief' select * from t4 where (id, name_id) in ((1, 1), (3, 3))",
+      "explain format='brief' select * from t5 where (id, name) in ((1, 'a'), (3, 'c'))"
+    ]
   }
 ]
