+++ planner/core/casetest/partition/testdata/integration_partition_suite_out.json
@@ -2386,5 +2386,48 @@
         ]
       }
     ]
+  },
+  {
+    "Name": "TestBatchPointGetPartitionForAccessObject",
+    "Cases": [
+      {
+        "SQL": "explain format='brief' select * from t1 where b in (1, 2)",
+        "Plan": [
+          "Batch_Point_Get 2.00 root table:t1, partition:p1,p2, index:b(b) keep order:false, desc:false"
+        ]
+      },
+      {
+        "SQL": "explain format='brief' select * from t1 where b in (1, 2, 1)",
+        "Plan": [
+          "Batch_Point_Get 3.00 root table:t1, partition:p1,p2, index:b(b) keep order:false, desc:false"
+        ]
+      },
+      {
+        "SQL": "explain format='brief' select * from t2 where id in (1, 3)",
+        "Plan": [
+          "Batch_Point_Get 2.00 root table:t2, partition:p0,p1 handle:[1 3], keep order:false, desc:false"
+        ]
+      },
+      {
+        "SQL": "explain format='brief' select * from t3 where id in (1, 3)",
+        "Plan": [
+          "Batch_Point_Get 2.00 root table:t3 handle:[1 3], keep order:false, desc:false"
+        ]
+      },
+      {
+        "SQL": "explain format='brief' select * from t4 where (id, name_id) in ((1, 1), (3, 3))",
+        "Plan": [
+          "IndexReader 2.00 root partition:p0,p1 index:IndexRangeScan",
+          "└─IndexRangeScan 2.00 cop[tikv] table:t4, index:id(id, name_id) range:[1 1,1 1], [3 3,3 3], keep order:false, stats:pseudo"
+        ]
+      },
+      {
+        "SQL": "explain format='brief' select * from t5 where (id, name) in ((1, 'a'), (3, 'c'))",
+        "Plan": [
+          "IndexReader 2.00 root partition:p0,p1 index:IndexRangeScan",
+          "└─IndexRangeScan 2.00 cop[tikv] table:t5, index:id(id, name) range:[1 \"a\",1 \"a\"], [3 \"c\",3 \"c\"], keep order:false, stats:pseudo"
+        ]
+      }
+    ]
   }
 ]
