+++ planner/core/point_get_plan_test.go
@@ -21,7 +21,6 @@ import (
 	"testing"
 	"time"
 
-	"github.com/pingcap/failpoint"
 	"github.com/pingcap/tidb/config"
 	"github.com/pingcap/tidb/metrics"
 	"github.com/pingcap/tidb/planner"
@@ -621,67 +620,6 @@ func TestBatchPointGetPartition(t *testing.T) {
 	))
 	tk.MustExec("delete from t where (a,b) in ((1,1),(2,2),(3,3),(4,4))")
 	tk.MustQuery("select * from t where (a, b) in ((1, 1), (2, 2), (3, 3), (4, 4))").Check(testkit.Rows())
-}
-
-func TestBatchPointGetPartitionForAccessObject(t *testing.T) {
-	failpoint.Enable("github.com/pingcap/tidb/planner/core/forceDynamicPrune", `return(true)`)
-	defer failpoint.Disable("github.com/pingcap/tidb/planner/core/forceDynamicPrune")
-	store := testkit.CreateMockStore(t)
-
-	tk := testkit.NewTestKit(t, store)
-	tk.MustExec("set @@tidb_partition_prune_mode = 'dynamic'")
-	tk.MustExec("use test")
-	tk.Session().GetSessionVars().EnableClusteredIndex = variable.ClusteredIndexDefModeOn
-	tk.MustExec("drop table if exists t")
-	tk.MustExec("create table t(a int, b int, UNIQUE KEY (b)) PARTITION BY HASH(b) PARTITIONS 4")
-	tk.MustExec("insert into t values(1, 1), (2, 2), (3, 3), (4, 4)")
-	tk.MustQuery("explain select * from t where b in (1, 2)").Check(testkit.Rows(
-		"Batch_Point_Get_1 2.00 root table:t, partition:p1,p2, index:b(b) keep order:false, desc:false"))
-	tk.MustQuery("explain select * from t where b in (1, 2, 1)").Check(testkit.Rows(
-		"Batch_Point_Get_1 3.00 root table:t, partition:p1,p2, index:b(b) keep order:false, desc:false"))
-
-	tk.MustExec("set @@session.tidb_enable_list_partition = ON")
-	tk.MustExec("drop table if exists t")
-	tk.MustExec("CREATE TABLE t (id int primary key, name_id int) PARTITION BY LIST(id) (" +
-		"partition p0 values IN (1, 2), " +
-		"partition p1 values IN (3, 4), " +
-		"partition p3 values IN (5))")
-	tk.MustExec("insert into t values(1, 1), (2, 2), (3, 3), (4, 4)")
-	tk.MustQuery("explain format='brief' select * from t where id in (1, 3)").Check(testkit.Rows(
-		"Batch_Point_Get 2.00 root table:t, partition:p0,p1 handle:[1 3], keep order:false, desc:false"))
-
-	tk.MustExec("set @@session.tidb_enable_list_partition = ON")
-	tk.MustExec("drop table if exists t0")
-	tk.MustExec("CREATE TABLE t0 (id int primary key, name_id int) PARTITION BY LIST COLUMNS(id) (" +
-		"partition p0 values IN (1, 2), " +
-		"partition p1 values IN (3, 4), " +
-		"partition p3 values IN (5))")
-	tk.MustExec("insert into t0 values(1, 1), (2, 2), (3, 3), (4, 4)")
-	tk.MustQuery("explain format='brief' select * from t0 where id in (1, 3)").Check(testkit.Rows(
-		"TableReader 2.00 root partition:p0,p1 data:TableRangeScan]\n" +
-			"[└─TableRangeScan 2.00 cop[tikv] table:t0 range:[1,1], [3,3], keep order:false, stats:pseudo"))
-
-	tk.MustExec("set @@session.tidb_enable_list_partition = ON")
-	tk.MustExec("drop table if exists t1")
-	tk.MustExec("CREATE TABLE t1 (id int, name_id int, unique key(id, name_id)) PARTITION BY LIST COLUMNS(id, name_id) (" +
-		"partition p0 values IN ((1, 1),(2, 2)), " +
-		"partition p1 values IN ((3, 3),(4, 4)), " +
-		"partition p3 values IN ((5, 5)))")
-	tk.MustExec("insert into t1 values(1, 1), (2, 2), (3, 3), (4, 4)")
-	tk.MustQuery("explain format='brief' select * from t1 where (id, name_id) in ((1, 1), (3, 3))").Check(testkit.Rows(
-		"IndexReader 2.00 root partition:p0,p1 index:IndexRangeScan]\n" +
-			"[└─IndexRangeScan 2.00 cop[tikv] table:t1, index:id(id, name_id) range:[1 1,1 1], [3 3,3 3], keep order:false, stats:pseudo"))
-
-	tk.MustExec("set @@session.tidb_enable_list_partition = ON")
-	tk.MustExec("drop table if exists t2")
-	tk.MustExec("CREATE TABLE t2 (id int, name varchar(10), unique key(id, name)) PARTITION BY LIST COLUMNS(id, name) (" +
-		"partition p0 values IN ((1,'a'),(2,'b')), " +
-		"partition p1 values IN ((3,'c'),(4,'d')), " +
-		"partition p3 values IN ((5,'e')))")
-	tk.MustExec("insert into t2 values(1, 'a'), (2, 'b'), (3, 'c'), (4, 'd')")
-	tk.MustQuery("explain format='brief' select * from t2 where (id, name) in ((1, 'a'), (3, 'c'))").Check(testkit.Rows(
-		"IndexReader 2.00 root partition:p0,p1 index:IndexRangeScan]\n" +
-			"[└─IndexRangeScan 2.00 cop[tikv] table:t2, index:id(id, name) range:[1 \"a\",1 \"a\"], [3 \"c\",3 \"c\"], keep order:false, stats:pseudo"))
 }
 
 func TestIssue19141(t *testing.T) {
